db-cbbr/cbbr_build/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    ./dev_python_packages.sh
#
attrs==22.2.0
    # via
    #   fiona
    #   pytest
certifi==2022.12.7
    # via
    #   fiona
    #   pyproj
click==8.1.3
    # via
    #   click-plugins
    #   cligj
    #   fiona
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
exceptiongroup==1.1.0
    # via pytest
fiona==1.9.1
    # via geopandas
future==0.18.3
    # via usaddress
geopandas==0.12.2
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
iniconfig==2.0.0
    # via pytest
munch==2.5.0
    # via fiona
numpy==1.24.2
    # via
    #   pandas
    #   shapely
packaging==23.0
    # via
    #   geopandas
    #   pytest
pandas==1.5.3
    # via
    #   -r requirements.in
    #   geopandas
pluggy==1.0.0
    # via pytest
probableparsing==0.0.1
    # via usaddress
psycopg2-binary==2.9.5
    # via -r requirements.in
pyproj==3.4.1
    # via geopandas
pytest==7.2.1
    # via -r requirements.in
python-crfsuite==0.9.9
    # via usaddress
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.21.1
    # via -r requirements.in
python-geosupport==1.0.8
    # via -r requirements.in
pytz==2022.7.1
    # via pandas
shapely==2.0.1
    # via geopandas
six==1.16.0
    # via
    #   munch
    #   python-dateutil
sqlalchemy==2.0.4
    # via -r requirements.in
tomli==2.0.1
    # via pytest
typing-extensions==4.5.0
    # via sqlalchemy
usaddress==0.5.10
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
